refactor: organize types and schemas into clean domain architecture
- Split monolithic types.ts into logical domain files:
  * core/ - abilities, damage types, conditions, stat blocks
  * character/ - characters, classes, equipment, spells
  * combat/ - actions, effects, encounters
  * creatures/ - NPCs and monsters
  * campaign/ - campaigns, party state, quests
  * logging/ - turn logs and roll logs
  * session/ - game sessions
  * versioning/ - migration support

- Split monolithic schemas.ts into corresponding schema files:
  * Matching domain structure for validation
  * Proper cross-references between domains
  * Clean parser function exports

- Benefits:
  * Single responsibility - each file has focused purpose
  * Logical grouping - related types co-located
  * Better maintainability and navigation
  * Tree shaking support for optimal bundles
  * Team collaboration friendly
  * Incremental development support

- Maintained backward compatibility:
  * All tests still pass (39/39)
  * Same public API through barrel exports
  * No breaking changes for consumers

- Added comprehensive README documenting:
  * Clean architecture principles
  * Domain organization rationale
  * Usage examples and import patterns
  * Migration support explanation
