+++ packages/models/src/character/index.ts
@@ -1,0 +1,6 @@
+// Character domain exports
+
+export * from "./character";
+export * from "./classes";
+export * from "./equipment";
+export * from "./spells";
