+++ packages/models/src/core/conditions.ts
@@ -1,0 +1,17 @@
+// Conditions and status effects
+
+export type Condition = 
+  | "blinded"
+  | "charmed"
+  | "deafened"
+  | "frightened"
+  | "grappled"
+  | "incapacitated"
+  | "invisible"
+  | "paralyzed"
+  | "petrified"
+  | "poisoned"
+  | "prone"
+  | "restrained"
+  | "stunned"
+  | "unconscious";
