+++ packages/models/src/core/index.ts
@@ -1,0 +1,6 @@
+// Core D&D mechanics exports
+
+export * from "./abilities";
+export * from "./damage";
+export * from "./conditions";
+export * from "./stats";
