+++ packages/models/src/creatures/index.ts
@@ -1,0 +1,4 @@
+// Creatures domain exports
+
+export * from "./npc";
+export * from "./monster";
