+++ packages/models/src/creatures/monster.ts
@@ -1,0 +1,12 @@
+// Monsters and challenging creatures
+
+import type { NPC } from "./npc";
+import type { Action } from "../combat/actions";
+
+export interface Monster extends NPC {
+  challengeRating: number; // 0-30
+  experiencePoints: number; // ≥ 0
+  legendaryActions?: Action[];
+  lairActions?: Action[];
+  regionalEffects?: string[];
+}
