+++ packages/models/src/index.ts
@@ -1,5 +1,14 @@
-// Export all types from the comprehensive domain model
-export * from "./types";
+// Core D&D mechanics
+export * from "./core";
 
-// Export all Zod schemas and parsers
+// Domain-specific exports
+export * from "./character";
+export * from "./combat";
+export * from "./creatures";
+export * from "./campaign";
+export * from "./logging";
+export * from "./session";
+export * from "./versioning";
+
+// Schema validation and parsers
 export * from "./schemas";
