+++ packages/models/src/versioning/index.ts
@@ -1,0 +1,3 @@
+// Versioning exports
+
+export * from "./versioned";
